requirements.txt
openai
python-dotenv
Flask
markdown
azure-ai-inference